Wider id values
- Increased id width from 4 to 8 bytes to avoid overflowing problems in long living systems.

+++ src/MvcTemplate.Components/Extensions/MvcGrid/MvcGridExtensions.cs
@@ -100,7 +100,7 @@ namespace MvcTemplate.Components.Extensions
             if (authorization == null)
                 return true;
 
-            Int32? account = context.HttpContext.User.Id();
+            Int64? account = context.HttpContext.User.Id();
             String? area = context.RouteData.Values["area"] as String;
             String? controller = context.RouteData.Values["controller"] as String;
