+++ src/MvcTemplate.Components/Extensions/MvcTree/MvcTree.cs
@@ -6,12 +6,12 @@ namespace MvcTemplate.Components.Extensions
     public class MvcTree
     {
         public List<MvcTreeNode> Nodes { get; set; }
-        public HashSet<Int32> SelectedIds { get; set; }
+        public HashSet<Int64> SelectedIds { get; set; }
 
         public MvcTree()
         {
             Nodes = new List<MvcTreeNode>();
-            SelectedIds = new HashSet<Int32>();
+            SelectedIds = new HashSet<Int64>();
         }
     }
 }
