+++ src/MvcTemplate.Components/Security/Authorization/IAuthorization.cs
@@ -4,7 +4,7 @@ namespace MvcTemplate.Components.Security
 {
     public interface IAuthorization
     {
-        Boolean IsGrantedFor(Int32? accountId, String? area, String? controller, String? action);
+        Boolean IsGrantedFor(Int64? accountId, String? area, String? controller, String? action);
 
         void Refresh();
     }
