+++ src/MvcTemplate.Objects/Models/Administration/Roles/RolePermission.cs
@@ -4,10 +4,10 @@ namespace MvcTemplate.Objects
 {
     public class RolePermission : BaseModel
     {
-        public Int32 RoleId { get; set; }
+        public Int64 RoleId { get; set; }
         public virtual Role Role { get; set; }
 
-        public Int32 PermissionId { get; set; }
+        public Int64 PermissionId { get; set; }
         public virtual Permission Permission { get; set; }
     }
 }
