+++ src/MvcTemplate.Objects/Views/Administration/Accounts/AccountEditView.cs
@@ -16,6 +16,6 @@ namespace MvcTemplate.Objects
 
         public Boolean IsLocked { get; set; }
 
-        public Int32? RoleId { get; set; }
+        public Int64? RoleId { get; set; }
     }
 }
