+++ src/MvcTemplate.Validators/Administration/Accounts/AccountValidator.cs
@@ -72,7 +72,7 @@ namespace MvcTemplate.Validators
             return isValid;
         }
 
-        private Boolean IsUniqueUsername(Int32 accountId, String? username)
+        private Boolean IsUniqueUsername(Int64 accountId, String? username)
         {
             Boolean isUnique = !UnitOfWork
                 .Select<Account>()
@@ -85,7 +85,7 @@ namespace MvcTemplate.Validators
 
             return isUnique;
         }
-        private Boolean IsUniqueEmail(Int32 accountId, String? email)
+        private Boolean IsUniqueEmail(Int64 accountId, String? email)
         {
             Boolean isUnique = !UnitOfWork
                 .Select<Account>()
@@ -114,7 +114,7 @@ namespace MvcTemplate.Validators
 
             return isCorrect;
         }
-        private Boolean IsCorrectPassword(Int32 accountId, String? password)
+        private Boolean IsCorrectPassword(Int64 accountId, String? password)
         {
             String passhash = UnitOfWork
                 .Select<Account>()
