+++ src/MvcTemplate.Validators/Administration/Roles/RoleValidator.cs
@@ -29,7 +29,7 @@ namespace MvcTemplate.Validators
             return isValid;
         }
 
-        private Boolean IsUniqueTitle(Int32 id, String? title)
+        private Boolean IsUniqueTitle(Int64 id, String? title)
         {
             Boolean isUnique = !UnitOfWork
                 .Select<Role>()
