+++ src/MvcTemplate.Validators/BaseValidator.cs
@@ -11,7 +11,7 @@ namespace MvcTemplate.Validators
     public abstract class BaseValidator : IValidator
     {
         public ModelStateDictionary ModelState { get; set; }
-        public Int32 CurrentAccountId { get; set; }
+        public Int64 CurrentAccountId { get; set; }
         public Alerts Alerts { get; set; }
 
         protected IUnitOfWork UnitOfWork { get; }
