+++ src/MvcTemplate.Validators/IValidator.cs
@@ -7,7 +7,7 @@ namespace MvcTemplate.Validators
     public interface IValidator : IDisposable
     {
         ModelStateDictionary ModelState { get; set; }
-        Int32 CurrentAccountId { get; set; }
+        Int64 CurrentAccountId { get; set; }
         Alerts Alerts { get; set; }
     }
 }
